+++ astropaint/profiles/Battaglia16.py
@@ -15,7 +15,7 @@ from astropy import units as u
 from astropy.constants import sigma_T, m_p
 from astropy.cosmology import Planck18_arXiv_v2 as cosmo
 
-from astropaint.lib.utilities import interpolate, LOS_integrate
+from astropaint.lib.utils import interpolate, LOS_integrate
 
 # ---------------Caching----------------
 # To cache templates use
